requirements.txt
tensorflow
numpy
Pillow
scipy